requirements.txt
# requirements.txt — VERSÃO FUNCIONAL
fastapi==0.121.1
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-settings==2.5.2
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
cachetools==5.5.0
requests==2.32.3
python-dotenv==1.2.1
email-validator==2.3.0
python-multipart==0.0.20
httpx==0.27.2
apscheduler==3.10.4
alembic==1.13.3
pytest
pytest-asyncio
httpx
slowapi